pyproject.toml
[project]
name = "django-fast-api-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "django>=4.2,<5.0",
    "django-cors-headers>=4.9.0",
    "fastapi>=0.135.1",
    "google-genai>=1.65.0",
    "google-generativeai>=0.8.6",
    "gunicorn>=25.1.0",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.41.0",
    "websockets>=16.0",
]
